refactor: move tsx files under view to the components folder
