+++ app/modal.tsx
@@ -1,7 +1,6 @@
 import { StatusBar } from "expo-status-bar";
 import { Platform, StyleSheet } from "react-native";
-import { Text, View } from "../components/Themed";
-import AddItem from "../view/AddItem";
+import { View, AddItem } from "../components";
 
 export default function ModalScreen() {
   return (
