+++ components/index.ts
@@ -1,0 +1,4 @@
+export { default as TodoList } from "./TodoList";
+export { default as AddItem } from "./AddItem";
+export * from "./Themed";
+export * from "./StyledText";
