+++ package.json
@@ -34,7 +34,9 @@
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
+    "@types/jest": "^29.5.8",
     "@types/react": "~18.2.14",
+    "@types/react-test-renderer": "^18.0.6",
     "jest": "^29.2.1",
     "jest-expo": "~49.0.0",
     "react-test-renderer": "18.2.0",
